requirements.txt
asgiref==3.6.0
Django==4.2
django-dotenv==1.4.2
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
psycopg==3.1.8
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
